19.11.21 chapte 06 input 추가(int형 변경)

+++ chapter_06/input.py
@@ -1,3 +1,3 @@
 b=input("정사각형 한변의 길이는?:")
-area=float(b)**2
+area=int(b)**2
 print("정사격의 넓이:{}".format(area))
